dedup validatePassword functionality
